# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Magooify, a self-contained Go application framework.",
        "title": "Magooify API",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/credits": {
            "get": {
                "description": "Returns the remaining OpenRouter credits for the account (when a management key is configured) and the session's accumulated spend in US dollars.",
                "produces": [
                    "application/json"
                ],
                "summary": "Get Credits",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/CreditsResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/health": {
            "get": {
                "description": "Returns operational status of the service.",
                "produces": [
                    "application/json"
                ],
                "summary": "Health Check",
                "responses": {
                    "200": {
                        "description": "Healthy response",
                        "schema": {
                            "$ref": "#/definitions/HealthResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/images": {
            "get": {
                "description": "Lists the processed images stored in the configured output directory, newest first.",
                "produces": [
                    "application/json"
                ],
                "summary": "List Stored Images",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ImagesResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/images/{filename}": {
            "get": {
                "description": "Serves a processed image stored in the output directory.",
                "produces": [
                    "application/octet-stream"
                ],
                "summary": "Get Stored Image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Image filename",
                        "name": "filename",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/info": {
            "get": {
                "description": "Returns system metrics, Go version, OS/Arch, and uptime.",
                "produces": [
                    "application/json"
                ],
                "summary": "System Info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SystemInfoResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/items": {
            "get": {
                "description": "Returns all stored items.",
                "produces": [
                    "application/json"
                ],
                "summary": "List Items",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ItemsResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new item.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create Item",
                "parameters": [
                    {
                        "description": "Item to create",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateItemRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/Item"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/models": {
            "get": {
                "description": "Lists OpenRouter models that accept image input with the estimated cost of processing a single image, so cheaper alternatives to the configured model are easy to spot. Exact per-image prices are used when published; otherwise costs are estimated from the per-token rates.",
                "produces": [
                    "application/json"
                ],
                "summary": "List Image-Capable Models",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ModelsResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/process": {
            "post": {
                "description": "Accepts an image captured by the camera or uploaded by the user, sends it to an OpenRouter vision model for processing, then stores the processed version of the image in the configured output directory.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Process an Image with OpenRouter",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Image to process",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Prompt to send with the image; defaults to the configured prompt",
                        "name": "prompt",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Optional filename to write the processed image to, replacing any existing file with that name",
                        "name": "output",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ProcessImageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/prompt": {
            "get": {
                "description": "Returns the prompt text sent to the model with each image.",
                "produces": [
                    "application/json"
                ],
                "summary": "Get Processing Prompt",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/user": {
            "get": {
                "description": "Returns authenticated user info based on environment or proxy headers.",
                "produces": [
                    "application/json"
                ],
                "summary": "Get Current User",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UserInfo"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "CreateItemRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Sample Item"
                }
            }
        },
        "CreditsResponse": {
            "type": "object",
            "properties": {
                "credits_available": {
                    "type": "boolean"
                },
                "remaining_credits": {
                    "type": "number"
                },
                "session_cost": {
                    "type": "number"
                },
                "total_credits": {
                    "type": "number"
                },
                "total_usage": {
                    "type": "number"
                }
            }
        },
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid request payload"
                }
            }
        },
        "HealthResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2026-08-04T12:00:00Z"
                }
            }
        },
        "ImagesResponse": {
            "type": "object",
            "properties": {
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/StoredImage"
                    }
                }
            }
        },
        "Item": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "ItemsResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Item"
                    }
                }
            }
        },
        "ModelInfo": {
            "type": "object",
            "properties": {
                "completion_per_million": {
                    "type": "number",
                    "example": 1.5
                },
                "estimated_image_cost": {
                    "type": "number",
                    "example": 0.0304
                },
                "id": {
                    "type": "string",
                    "example": "google/gemini-3.1-flash-lite-image"
                },
                "input_image_cost": {
                    "type": "number",
                    "example": 0.0004
                },
                "input_image_cost_known": {
                    "type": "boolean",
                    "example": false
                },
                "name": {
                    "type": "string",
                    "example": "Google: Nano Banana 2 Lite (Gemini 3.1 Flash Lite Image)"
                },
                "output_image_cost": {
                    "type": "number",
                    "example": 0.03
                },
                "output_image_cost_known": {
                    "type": "boolean",
                    "example": true
                },
                "outputs_images": {
                    "type": "boolean",
                    "example": true
                },
                "prompt_per_million": {
                    "type": "number",
                    "example": 0.25
                },
                "request_cost": {
                    "type": "number",
                    "example": 0
                }
            }
        },
        "ModelsResponse": {
            "type": "object",
            "properties": {
                "models": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ModelInfo"
                    }
                }
            }
        },
        "ProcessImageResponse": {
            "type": "object",
            "properties": {
                "cost": {
                    "type": "number"
                },
                "filename": {
                    "type": "string",
                    "example": "img-20260806-123000-a1b2c3d4.jpg"
                },
                "model": {
                    "type": "string",
                    "example": "google/gemini-3.1-flash-lite-image"
                },
                "processed_at": {
                    "type": "string"
                },
                "session_cost": {
                    "type": "number"
                }
            }
        },
        "StoredImage": {
            "type": "object",
            "properties": {
                "filename": {
                    "type": "string",
                    "example": "img-20260806-123000-a1b2c3d4.jpg"
                },
                "modified": {
                    "type": "string"
                },
                "size": {
                    "type": "integer",
                    "example": 48291
                }
            }
        },
        "SystemInfoResponse": {
            "type": "object",
            "properties": {
                "arch": {
                    "type": "string",
                    "example": "amd64"
                },
                "go_version": {
                    "type": "string",
                    "example": "go1.24.4"
                },
                "mode": {
                    "type": "string",
                    "example": "desktop"
                },
                "os": {
                    "type": "string",
                    "example": "linux"
                },
                "uptime": {
                    "type": "string",
                    "example": "12m30s"
                }
            }
        },
        "UserInfo": {
            "type": "object",
            "properties": {
                "auth_type": {
                    "type": "string",
                    "example": "local_env"
                },
                "mode": {
                    "type": "string",
                    "example": "desktop"
                },
                "username": {
                    "type": "string",
                    "example": "alice"
                }
            }
        }
    }
}